requirements.txt
python-dotenv
smolagents